refactor: extract NewModal into its own folder
